Refactor Routes
